#50: Refactor to rd_token.py.
