+++ src/raindrop_todoist_syncer/main.py
@@ -4,9 +4,9 @@ from loguru import logger
 
 from raindrop_todoist_syncer.db_manage import DatabaseManager
 from raindrop_todoist_syncer.rd_process import RaindropsProcessor
+from raindrop_todoist_syncer.rd_token import RaindropAccessTokenRefresher
 from raindrop_todoist_syncer.raindrop import (
     EnvironmentVariablesFileManager,
-    RaindropAccessTokenRefresher,
     RaindropClient,
     RaindropCredentialsManager,
 )
