+++ tests/conftest.py
@@ -8,10 +8,10 @@ import pytest
 from requests import HTTPError
 
 from raindrop_todoist_syncer.raindrop import (
-    RaindropAccessTokenRefresher,
     RaindropCredentialsManager,
     EnvironmentVariablesFileManager,
 )
+from raindrop_todoist_syncer.rd_token import RaindropAccessTokenRefresher
 from raindrop_todoist_syncer.rd_class import Raindrop
 
 # ---------------------------------- env vars-------------------------------------------
