+++ tests/integration/test_main.py
@@ -92,7 +92,7 @@ class TestStaleTokenFunctionality:
         caplog,
     ):
         with patch(
-            "raindrop.RaindropAccessTokenRefresher.refresh_token_process_runner"
+            "raindrop_todoist_syncer.rd_token.RaindropAccessTokenRefresher.refresh_token_process_runner"
         ) as mock_refresh_token_runner, patch(
             "raindrop.RaindropClient.stale_token", return_value=True
         ) as mock_valid_token:  # noqa
@@ -108,7 +108,7 @@ class TestStaleTokenFunctionality:
         caplog,
     ):
         with patch(
-            "raindrop.RaindropAccessTokenRefresher.refresh_token_process_runner"
+            "raindrop_todoist_syncer.rd_token.RaindropAccessTokenRefresher.refresh_token_process_runner"
         ) as mock_refresh_token_runner, patch(
             "raindrop.RaindropClient.stale_token", return_value=False
         ) as mock_valid_token:  # noqa
